fix: :truck: Moved wall component to utils
Issue: #39
